requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
cx-Oracle==8.3.0
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
idna==3.4
mariadb==1.1.8
mysql-connector-python==8.2.0
packaging==23.2
protobuf==4.21.12
pydantic==2.4.2
pydantic_core==2.10.1
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
